pyproject.toml
[project]
name = "heart"
version = "0.2.0"
classifiers = [
  "Development Status :: 3 - Alpha",

  # Indicate who your project is intended for
  "Intended Audience :: Developers",
  "Intended Audience :: Education",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Music Lovers",
  "Intended Audience :: Artists",
  "Intended Audience :: Dancers",
  "Intended Audience :: Other Audience",
  "Environment :: MacOS X",
  "Environment :: Linux",
  "Environment :: Raspberry Pi",
  "Topic :: Software Development :: Libraries :: pygame",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
]
keywords = []
authors = []
requires-python = ">= 3.11"

dependencies = [
    "bleak",
    "numba",
    "numpy",
    "openant",
    "opencv-python",
    "Pillow",
    "pygame==2.6.1",
    "pyserial",
    "reactivex",
    "requests",
    "scipy",
    "toml",
    "tqdm",
    "typer-slim",
    "PyOpenGL>=3.1.0",
    "moderngl",
    "sounddevice",
    "paho-mqtt",
]

[project.optional-dependencies]
dev = [
    "Pillow",
    "scipy",
    "mypy",
    "types-requests",
    "types-toml",
    # Git
    "pre-commit",
]

[build-system]
requires = ["uv_build>=0.9.6,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
heart = { path = "./src/heart" }


[tool.setuptools.package-data]
heart = ["*.png"]

[project.scripts]
totem = "heart.loop:main"
totem_debug = "heart.x.cli:main"

[tool.docformatter]
recursive = true
blank = true

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["E501", "F403", "F405"]

[tool.ruff.lint.per-file-ignores]
"src/heart/display/shaders/shader.py" = ["F403", "F405"]
"src/heart/display/renderers/three_fractal.py" = ["F403", "F405"]


[tool.mypy]
python_version = "3.11"
mypy_path = "src"
packages = [
    "heart.device",
    "heart.environment",
    "heart.loop",
    "heart.peripheral",
    "heart.utilities",
]
warn_unused_configs = true
warn_unused_ignores = true
check_untyped_defs = true
color_output = false
error_summary = false
show_error_context = false
pretty = false

[[tool.mypy.overrides]]
module = [
    "bleak",
    "bleak.*",
    "cv2",
    "cv2.*",
    "moderngl",
    "numpy",
    "numpy.*",
    "openant",
    "openant.*",
    "PIL",
    "PIL.*",
    "pygame",
    "pygame.*",
    "requests",
    "rgbmatrix",
    "serial", 
    "serial.*", 
    "bluezero", 
    "bluezero.*", 
    "typer", 
    "toml", 
    "sounddevice", 
    "usb", 
    "usb.*", 
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "heart.assets",
    "heart.assets.*",
    "heart.display",
    "heart.display.*",
    "heart.navigation",
    "heart.navigation.*",
]
follow_imports = "skip"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-vv -n auto --dist=loadscope"
markers = [
    "timeout(duration): mark tests that should fail if they exceed the given duration",
]

[dependency-groups]
dev = [
  {include-group = "lint"},
  {include-group = "test"}
]
test = [
    "pytest>=8.4.2",
    "pytest-benchmark>=5.2.0",
    "pytest-xdist>=3.8.0",
]
lint = [
    "docformatter[tomli]>=1.7.7",
    "isort>=7.0.0",
    "mdformat>=1.0.0",
    "mypy>=1.13.0",
    "ruff>=0.14.3",
]
